Fix removal of Fixnum in ruby-next in number formatter
The current ruby head removed Fixnum and Bignum classes, which has made
the build break. I also don't believe the Bignum support was allowing
the formatting of arbitrary bignums anyway.

Fix it by attempting to coalesce the passed-in number to an int64_t, and
if that fails, falling back to calling unum_format_decimal.