Add #as_skeleton and #set_date_format_from_skeleton to DateTimeFormatter
As documented in the tests, the main purpose of this functionality is to
allow:
    * Loading up one of the standard date/time patterns :long, :short,
      etc
    * Converting it into a skeleton
    * Manipulating the skeleton somehow
    * Converting the skeleton back into a real pattern
    * Formatting a time with it

The use-case in question I have is being able to modify the standard
skeleton patterns to use `j` for locale-appropriate 12/24 hour time,
instead of H/h being specified per-locale. This means that locales with
hours keywords like "en_AU@hours=h23" can be made to produce 24 hour
time, which ICU would not normally do.